Refactor comportamientos legacy and add API server
