+++ requirements.txt
@@ -1,0 +1,2 @@
+fastapi==0.110.0
+uvicorn==0.27.0
