+++ src/game/combate/ia/ia_comportamiento.py
@@ -1,4 +1,4 @@
 """Compatibilidad: expone BehaviorExecutor para IA"""
-from src.game.comportamientos.legacy.OLD.behavior_executor import BehaviorExecutor
+from src.game.comportamientos.legacy.behavior_executor import BehaviorExecutor
 
 __all__ = ["BehaviorExecutor"]
