+++ src/game/comportamientos/legacy/__init__.py
@@ -1,9 +1,9 @@
 """Compatibilidad con el antiguo sistema de comportamientos."""
 
-from .OLD.behavior_executor import BehaviorExecutor
-from .OLD.movement_behaviors import MovementBehavior, MovementProcessor
-from .OLD.combat_behaviors import CombatBehavior, CombatProcessor
-from .OLD.behavior_restrictions import BehaviorRestrictions
+from .behavior_executor import BehaviorExecutor
+from .movement_behaviors import MovementBehavior, MovementProcessor
+from .combat_behaviors import CombatBehavior, CombatProcessor
+from .behavior_restrictions import BehaviorRestrictions
 
 __all__ = [
     "BehaviorExecutor",
