+++ src/game/comportamientos/legacy/behavior_component.py
