+++ src/game/comportamientos/legacy/behavior_restrictions.py
