+++ src/game/comportamientos/legacy/behavior_utils.py
