+++ src/game/comportamientos/legacy/combat_behaviors.py
