+++ src/game/comportamientos/legacy/movement_behaviors.py
